rename useDataChannel hook file
