+++ packages/react/src/hooks/index.ts
@@ -1,7 +1,7 @@
 export * from './participant-hooks';
 export * from './token-hooks';
 export * from './track-hooks';
-export * from './dataChannel-hooks';
+export { useDataChannel } from './useDataChannel';
 export { useTracks } from './useTracks';
 export { useGridLayout } from './useGridLayout';
 export { useVisualStableUpdate } from './useVisualStableUpdate';
